% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{metadata}
\alias{metadata}
\title{Metadata Extractor for OpenAI Reports}
\usage{
metadata(file_path)
}
\arguments{
\item{file_path}{the OpenAI .jsonl Evaluation file path}
}
\value{
A dataframe consisiting of the data_source_idx and the original dataset uploaded to OpenAI
}
\description{
This function extracts item.* fields and data_source_idx from an OpenAI evaluation
export, allowing users to reconnect model results to their original dataset using
the unique identifier extracted using the results() function.
}
\examples{
\dontrun{
metadata("data/example.jsonl")
}
}
