% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qwetAC1.R
\name{qwetAC1}
\alias{qwetAC1}
\title{Calcuate Qwet's AC1 for OpenAI Model Outputs}
\usage{
qwetAC1(df, groundtruth)
}
\arguments{
\item{df}{the dataframe created by this package's the results function}

\item{groundtruth}{The name of the column containing the ground_truth labels.}
}
\value{
a list consisting of the Cohen's kappa between each model run and the ground truth
}
\description{
This function computes a Qwet's AC1, an alternate statistic for inter-rater reliability,
between a specified ground truth column and each model's output column in a dataframe.
}
\examples{
qwetAC1(data, "groundtruth")
}
