% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{results}
\alias{results}
\title{Extract Model Results from OpenAI Evaluation reports}
\usage{
results(file_path, groundtruth)
}
\arguments{
\item{file_path}{the OpenAI .jsonl Evaluation file path}

\item{groundtruth}{the full name of the ground-truth column}
}
\value{
A dataframe consisiting of the data_source_idx, the ground-truth column,
and any model-result pairing for each model in the test.
}
\description{
This function will read a JSONL file containing model-generated outputs,
extract relevant fields, and reshape the data so that outputs from multiple
models for each unique item are displayed side by side in a standard dataframe.
}
\examples{
\dontrun{
results("data/example.jsonl", "groundtruth")
}
}
